pyproject.toml
[project]
name = "papyrix-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "freetype-py>=2.5.1",
    "pillow>=12.0.0",
]

[project.urls]
Homepage = "https://github.com/bigbag/papyrix-reader"
Repository = "https://github.com/bigbag/papyrix-reader"
Issues = "https://github.com/bigbag/papyrix-reader/issues"
